Add Nix language server support (#6)
This adds very basic support for the
[nixd](https://github.com/nix-community/nixd) language server, which is
the most popular Nix language server.

Note that currently, the logic to retrieve the language server is very
minimal, to say the least. It will simply complain if it doesn't find
the binary in the PATH. However, we cannot download it, as some other
extensions would probably do, due to the source-aware nature of Nix and
nixpkgs. Let me know if you have better ideas here.

It also bumps the schema files to what's upstream in
[tree-sitter-nix](https://github.com/nix-community/tree-sitter-nix).

Closes #1, related to
https://github.com/zed-industries/extensions/pull/857

---------

Signed-off-by: Moritz Sanft <58110325+msanft@users.noreply.github.com>
Co-authored-by: Marshall Bowers <elliott.codes@gmail.com>